+++ changelog.md
@@ -3,6 +3,7 @@
 v1.16.0:
 
 - Add initial compatibility with WoW 12.x / Midnight. Please see readme/in-game FAQ page for details about this significant change.
+- Add option to forcibly disable cooldown numbers on TargetInfo buff and debuff frames.
 
 v1.15.23: